+++ react/DisplayButtonTypes.js
@@ -16,11 +16,11 @@ const displayButtonTypes = {
 }
 
 export function getDisplayButtonNames() {
-  return map(opt => opt.name ,values(displayButtonTypes))
+  return map(opt => opt.name, values(displayButtonTypes))
 }
 
 export function getDisplayButtonValues() {
-  return map(opt => opt.value ,values(displayButtonTypes))
+  return map(opt => opt.value, values(displayButtonTypes))
 }
 
 export default displayButtonTypes