+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
-- Add buy button display option on StoreFront
+- Add buy button display option on StoreFront.
 
 ## [2.1.2] - 2018-12-12
 ### Fixed