+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.2.0] - 2018-12-14
+### Changed
+- Add buy button display option on StoreFront.
+
 ## [2.1.2] - 2018-12-12
 ### Fixed
 - Unavailable button layout.